third_party/WebKit/LayoutTests/inspector-protocol/runtime/runtime-console-timestamp.html

<html>
<head>
<script type="text/javascript" src="../../http/tests/inspector-protocol/resources/inspector-protocol-test.js"></script>
<script>

function test()
{

    var messages = [];

    function messageAdded(data)
    {
        var payload = data.params;

        if (messages.length > 0)
            InspectorTest.log("Message " + messages.length + " has non-decreasing timestamp: " + (payload.timestamp >= messages[messages.length - 1].timestamp));

        messages.push(payload);
        InspectorTest.log("Message has timestamp: " + !!payload.timestamp);

        InspectorTest.log("Message timestamp doesn't differ too much from current time (one minute interval): " + (Math.abs(new Date().getTime() - payload.timestamp) < 60000));
        if (messages.length === 3)
            InspectorTest.completeTest();
    }

    InspectorTest.eventHandler["Runtime.consoleAPICalled"] = messageAdded;
    InspectorTest.sendCommand("Runtime.enable", {});

    InspectorTest.sendCommand("Runtime.evaluate", { expression: "console.log('testUnique'); for (var i = 0; i < 2; ++i) console.log('testDouble');" });
}

</script>
</head>
<body onload="runTest()">
</body>
</html>
