third_party/WebKit/LayoutTests/inspector-protocol/runtime/runtime-console-timestamp-expected.txt

CONSOLE MESSAGE: line 1: testUnique
CONSOLE MESSAGE: line 1: testDouble
CONSOLE MESSAGE: line 1: testDouble
Message has timestamp: true
Message timestamp doesn't differ too much from current time (one minute interval): true
Message 1 has non-decreasing timestamp: true
Message has timestamp: true
Message timestamp doesn't differ too much from current time (one minute interval): true
Message 2 has non-decreasing timestamp: true
Message has timestamp: true
Message timestamp doesn't differ too much from current time (one minute interval): true
